Refractoring signIn, signUp, GetCurrentUser useCase
