+++ app/core/domain/repositories/AuthRepository.ts
@@ -1,0 +1,8 @@
+import { User } from '@/core/domain/entities/User';
+
+export interface AuthRepository {
+    signIn(email: string, password: string): Promise<User>;
+    signUp(email: string, password: string, displayName?: string): Promise<User>;
+    signOut(): Promise<void>;
+    getCurrentUser(): Promise<User | null>;
+}
