+++ app/core/domain/usecases/user/GetCurrentUserUseCase.ts
@@ -1,0 +1,9 @@
+import { AuthRepository} from "@/core/domain/repositories/AuthRepository.ts";
+import { User } from "@/core/domain/entities/User.ts";
+export class GetCurrentUserUseCase {
+    constructor(private authRepository: AuthRepository) {}
+
+    async execute(): Promise<User | null> {
+        return this.authRepository.getCurrentUser();
+    }
+}
