+++ app/core/domain/usecases/user/SignInUseCase.ts
@@ -1,0 +1,9 @@
+import { AuthRepository} from "@/core/domain/repositories/AuthRepository.ts";
+import { User } from "@/core/domain/entities/User.ts";
+export class SignInUseCase {
+    constructor(private authRepository: AuthRepository) {}
+
+    async execute(email: string, password: string): Promise<User> {
+        return this.authRepository.signIn(email, password);
+    }
+}
