+++ app/core/domain/usecases/user/SignOutUseCase.ts
@@ -1,0 +1,9 @@
+import { AuthRepository} from "@/core/domain/repositories/AuthRepository.ts";
+import { User } from "@/core/domain/entities/User.ts";
+export class SignOutUseCase {
+    constructor(private authRepository: AuthRepository) {}
+
+    async execute(): Promise<void> {
+        return this.authRepository.signOut();
+    }
+}
