+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@fontsource/roboto": "^5.2.5",
     "@react-native-community/datetimepicker": "^8.3.0",
+    "@react-native-community/geolocation": "^3.4.0",
     "@react-native-firebase/app": "^22.2.0",
     "@react-native-firebase/auth": "^22.2.0",
     "@react-native-firebase/firestore": "^22.2.0",
@@ -20,6 +21,7 @@
     "@react-navigation/native-stack": "^7.3.13",
     "@react-navigation/stack": "^7.3.1",
     "@reduxjs/toolkit": "^2.8.1",
+    "axios": "^1.9.0",
     "date-fns": "^4.1.0",
     "expo-localization": "^16.1.5",
     "formik": "^2.4.6",
@@ -27,6 +29,8 @@
     "lodash.debounce": "^4.0.8",
     "react": "19.0.0",
     "react-native": "0.79.2",
+    "react-native-geocoding": "^0.5.0",
+    "react-native-geolocation-service": "^5.3.1",
     "react-native-gesture-handler": "^2.25.0",
     "react-native-modal-datetime-picker": "^18.0.0",
     "react-native-reanimated": "^3.17.5",
